# internal/handler/openapi.yaml
openapi: 3.0.3
info:
  title: Ticket Reservation API
  version: 0.3.0
  description: |
    Seat reservation with temporary holds.

    A seat is claimed with a hold that expires after five minutes. Only the user
    holding a seat may confirm or release it, and a hold that has run out frees
    the seat immediately, whether or not the background sweeper has noticed yet.

    **Live updates**: `GET /events/{eventID}/stream` keeps a connection open and
    pushes a notice whenever a seat changes, so a client does not have to poll.

    **Limits**: every caller has an allowance. Signing in is limited far more
    tightly than reading, because each attempt costs a password hash. A refusal is
    a 429 carrying `Retry-After`.

    **Retries**: every write endpoint accepts an `Idempotency-Key` header. A
    repeated request carrying the same key gets the first attempt's answer back,
    marked with `Idempotent-Replay: true`, rather than being executed twice.

    **Identity**: writing endpoints need a bearer token; reading ones do not. The
    caller is whoever the token names, so a client can no longer act as anyone it
    likes by asserting a header.
servers:
  - url: /
    description: The server this page was loaded from
security:
  - BearerToken: []
...
paths:
...
  /auth/register:
    post:
      tags: [accounts]
      summary: Create an account
      description: |
        The address is stored lowered and trimmed, so signing up as Dilara@x.com
        and signing in as dilara@x.com reach the same account.

        Passwords are between 8 characters and 72 bytes. The upper bound is not
        arbitrary: bcrypt reads only the first 72 bytes, and accepting a longer
        password would silently ignore the rest of it.

        No token comes back. Signing in is a separate call.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "201":
          description: The account exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: The address already belongs to an account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: email_taken
                  message: that email address is already registered
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/login:
    post:
      tags: [accounts]
      summary: Sign in and get a token
      description: |
        Returns a bearer token for the account. Send it as
        `Authorization: Bearer <token>` on the writing endpoints.

        A wrong password and an unknown address give the same answer, and take a
        comparable amount of time, so that this endpoint cannot be used to find
        out which addresses are registered.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: The address or the password is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: invalid_credentials
                  message: the email address or password is wrong
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/verify:
    post:
      tags: [accounts]
      summary: Confirm an email address
      description: |
        Follows the link sent on registration. The token works once and stops
        working after a day.

        A POST rather than a GET on the link itself, because a mailbox scanner
        or a link preview fetching a GET would spend a single-use token before
        the person ever clicked it. The page behind the link reads the token out
        of the query string and sends it here.

        Until an address is confirmed nothing is emailed to it. Anybody can
        register with somebody else's address, and sending to it would make this
        service a way to post mail to strangers.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        "204":
          description: The address is confirmed
        "400":
          $ref: "#/components/responses/BadRequest"
        "410":
          description: |
            Unknown, already spent, or expired. One answer for all three:
            telling them apart would say whether a token was ever real.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: verification_not_usable
                  message: that verification link is no longer valid
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/verify/resend:
    post:
      tags: [accounts]
      summary: Send the confirmation link again
      description: |
        Needs a token, so it can only ask for the caller's own address. An
        endpoint that took an address would send mail to anybody who was named.

        Earlier links are left working: somebody who clicks the first message
        after asking for a second should not be told their link is broken.

        Answers 202 rather than 204, because the message is queued rather than
        already gone.
      responses:
        "202":
          description: A fresh link has been queued
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: The address is already confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: already_verified
                  message: the address is already verified
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/logout:
    post:
      tags: [accounts]
      summary: Sign out
      description: |
        Refuses the caller's own token for whatever is left of its life.

        It takes no body and names no token: the one being signed out is the one
        the request arrived with. A caller that could name a token could sign out
        somebody else's.

        Only that token is affected. Signing in twice gives two tokens, and this
        ends one of them, which is what makes signing out of one device possible.
        The record is kept only until the token would have expired anyway.
      responses:
        "204":
          description: The token will no longer be accepted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "501":
          description: |
            This deployment has nothing wired in to remember a signed out token.
            Not a failure: the service would otherwise be reporting that it had
            done something it cannot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: logout_unavailable
                  message: signing out is not available
...
  /events/{eventID}:
    get:
      tags: [events]
      summary: Get one event
      description: |
        Everything the detail page shows, including how many seats are still
        open. The counts are made here rather than left to the caller, because
        counting from outside means fetching every seat in the hall.
      security: []
      parameters:
        - $ref: "#/components/parameters/EventID"
      responses:
        "200":
          description: The event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /events/{eventID}/seats:
    get:
      tags: [events]
      summary: Get the seat map of an event
      description: |
        Seats are ordered by row and then number. A seat whose hold has expired is
        reported as `available` even if the sweeper has not cleaned it up yet, and
        the identity of whoever holds a seat is never exposed here.
      security: []
      parameters:
        - $ref: "#/components/parameters/EventID"
      responses:
        "200":
          description: The seat map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeatMap"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /events/{eventID}/stream:
    get:
      tags: [events]
      summary: Watch an event for changes
      description: |
        A server-sent event stream that stays open and pushes a notice whenever
        something changes, so a client does not have to keep asking.

        Each notice says *what* changed rather than what it changed to, so the
        client fetches the seat map when it hears one. That keeps a notice from
        going stale in flight and from carrying more than the seat map endpoint
        would.

        Kinds sent:

        - `seat_changed` — a seat's status is no longer what you last saw. Public,
          so every watcher of the event gets it.
        - `event_cancelled` — this event has been withdrawn from sale. Public,
          so every watcher gets it; the people who hold tickets also get a stored
          notice they will find whenever they next look.
        - `turn_came` — a seat you were waiting for is now held for you. It
          carries the `holdId` and the `expiresAt` of the hold that was made, so
          the client can confirm it without asking. Sent only to the caller it
          concerns, so this stream has to be opened with a token; `EventSource`
          cannot send one, which is why the bundled page reads it with `fetch`.

        Comment lines are sent every 25 seconds to keep proxies from closing an
        idle connection. Browsers reconnect on their own when it drops.
      security: []
      parameters:
        - $ref: "#/components/parameters/EventID"
      responses:
        "200":
          description: The stream is open
          content:
            text/event-stream:
              schema:
                type: string
              example: |
                event: seat_changed
                data: {"kind":"seat_changed","eventId":"event-1","seatId":"A1","at":"2026-08-18T12:00:00Z"}
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          description: The server cannot hold a connection open
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: stream_unavailable
                  message: live updates are not available
  /admin/events:
    post:
      tags: [admin]
      summary: Create an event
      description: |
        Requires an administrator. The role is read from storage on every
        request rather than carried in the token, so taking it away takes effect
        at once.

        The id is assigned here rather than taken from the caller, so two people
        stocking the service at the same time cannot collide over a name they
        both thought was free. It comes back in the body and in `Location`.

        There is no endpoint that makes an administrator. The first one is made
        with the `admin` command, which needs database access — an endpoint that
        can create an administrator is an endpoint that can be talked into
        creating one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewEvent"
      responses:
        "201":
          description: The event, with the id it was given
          headers:
            Location:
              schema:
                type: string
              description: Where the new event can be read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: An event with that id already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: event_exists
                  message: an event with that id already exists
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /admin/events/{eventID}:
    patch:
      tags: [admin]
      summary: Change an event
      description: |
        Every field is optional: a caller fixing a typo in the venue should not
        have to repeat the name and the start time, and repeating them is how one
        of them gets sent back stale.

        A cancelled event cannot be edited. It is a record of something that is
        not happening, and editing it would quietly turn it into a different one.
      parameters:
        - $ref: "#/components/parameters/EventID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventChanges"
      responses:
        "200":
          description: The event as it now stands
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The event is cancelled, so it is no longer editable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: event_already_cancelled
                  message: the event is already cancelled

    delete:
      tags: [admin]
      summary: Remove an event outright
      description: |
        Refused while any of the event's seats is held or sold. Removing a
        mistake is one thing; removing something people have tickets to is
        another, and cancelling is what that is for.

        The check and the delete are one statement, so a seat that stops being
        available in between cannot slip through.
      parameters:
        - $ref: "#/components/parameters/EventID"
      responses:
        "204":
          description: The event is gone
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Seats of this event are held or sold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: event_in_use
                  message: the event has seats that are held or sold
  /admin/events/{eventID}/cancel:
    post:
      tags: [admin]
      summary: Withdraw an event from sale
      description: |
        Marks the event rather than removing it, so the people holding tickets to
        it can still see what happened. It stays in the catalogue with
        `cancelled: true`.

        No new seats may be claimed afterwards. Anyone who already holds one may
        still confirm it: they were part way through a purchase when this
        happened, and taking it away from them is a worse answer than letting
        them finish.

        Takes no body. There is one thing to say and the path says which event.
      parameters:
        - $ref: "#/components/parameters/EventID"
      responses:
        "200":
          description: The event, now withdrawn
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: It was already withdrawn
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: event_already_cancelled
                  message: the event is already cancelled
  /admin/events/{eventID}/seats:
    post:
      tags: [admin]
      summary: Add rows of seats to an event
      description: |
        Requires an administrator. Builds the given rows, each numbered from one
        to `seatsPerRow`, so `rows: ["A","B"]` with `seatsPerRow: 3` makes A1, A2,
        A3, B1, B2 and B3.

        Seats that already exist are left alone, so adding a row twice does not
        disturb one that is already held or sold. At most 2000 seats in one
        request: without a limit one mistyped number becomes millions of rows.
      parameters:
        - $ref: "#/components/parameters/EventID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewSeats"
      responses:
        "201":
          description: How many seats were built
          content:
            application/json:
              schema:
                type: object
                required: [eventId, created]
                properties:
                  eventId:
                    type: string
                    example: event-1
                  created:
                    type: integer
                    example: 20
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /events/{eventID}/seats/{seatID}/hold:
    post:
      tags: [holds]
      summary: Hold a seat
      description: |
        Claims the seat for the caller for five minutes. Out of any number of
        simultaneous claims on one seat, exactly one succeeds; the others get 409.
      parameters:
        - $ref: "#/components/parameters/EventID"
        - $ref: "#/components/parameters/SeatID"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: The seat is held
          headers:
            Location:
              description: Where the new hold can be addressed
              schema:
                type: string
                example: /holds/MISGMDTLRSNOFSQWK7APCS3FAZ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hold"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: |
            The seat is already held by a live hold or already reserved
            (`seat_not_available`), or an earlier request with the same
            Idempotency-Key is still running (`idempotency_in_progress`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: seat_not_available
                  message: seat is not available
        "422":
          $ref: "#/components/responses/KeyReused"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /notifications:
    get:
      tags: [notifications]
      summary: What the caller has not been told yet
      description: |
        The caller's notices, newest first, read and unread together. `unread` is
        given separately so a page can show a count without walking the list,
        which is the one thing it needs before anybody opens it.

        Stored rather than only pushed down the stream, because the point of
        being told is that you were not looking: a notice that only reaches
        people who happen to have the page open is not a notification. The event
        name is carried in each notice, since the event may have been removed by
        the time anyone reads it.
      responses:
        "200":
          description: The caller's notices
          content:
            application/json:
              schema:
                type: object
                required: [unread, notifications]
                properties:
                  unread:
                    type: integer
                    example: 1
                  notifications:
                    type: array
                    items:
                      $ref: "#/components/schemas/Notification"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      description: |
        A signed token naming the caller, sent as `Authorization: Bearer <token>`.

        The payload is signed, not encrypted: anyone holding a token can read the
        user id out of it, but nobody can change it without the signing secret.

        Mint one for local use with `go run ./cmd/token dilara`, then paste it
        into **Authorize** above.
...
  responses:
...
    BodyTooLarge:
      description: The request body exceeds the 8 KiB limit
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: request_body_too_large
              message: the request body is too large
...